fix code smell add provider
